qibla-api/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
geopy==2.4.1
gunicorn==21.2.0
wmm-calculator>=0.1.0
numpy>=1.24